LayoutTests/storage/statement-error-callback.html

<html>
<head>
<script>

function log(message)
{
    document.getElementById("console").innerHTML += message + "<br>";
}

function finishTest()
{
    log("Test Complete");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

var txCallbackCount = 0;
var NUMBER_OF_TRANSACTIONS = 7;
var database;

function transactionErrorFunction(error)
{
    log("PASS - the transaction error callback was invoked.");
    if (++txCallbackCount == NUMBER_OF_TRANSACTIONS)
        finishTest();
}

function transactionSuccessFunction(message)
{
    log("FAIL - the transaction success callback should not be invoked.");
    if (++txCallbackCount == NUMBER_OF_TRANSACTIONS)
        finishTest();
}

function runTransactionExpectedToFail(statementErrorCallback)
{
    database.transaction(function(tx) {
        tx.executeSql("CREATE TABLE IF NOT EXISTS StatementErrorCallbackTest (randomData)");
        tx.executeSql("INSERT INTO StatementErrorCallbackTest (randomData) VALUES (?)", ['test']);
        tx.executeSql("THIS STATEMENT WILL FAIL", [],
            function(tx, data) {
                log("FAIL - this statement should have failed");
                finishTest();
            }, statementErrorCallback);
        tx.executeSql("INSERT INTO StatementErrorCallbackTest (randomData) VALUES (?)", ['test1'],
            function(error) { log("FAIL - This statement should not have been executed"); },
            function() { log("FAIL - This statement should not have been executed"); });
    }, transactionErrorFunction, transactionSuccessFunction);
}

function runTest()
{
    if (window.layoutTestController) {
        layoutTestController.clearAllDatabases();
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
    }

    database = openDatabase("bug-28872", "1.0", "statement error callback test", 1024);
    database.transaction(function(tx) {
        tx.executeSql("CREATE TABLE IF NOT EXISTS StatementErrorCallbackTest (randomData)");
        tx.executeSql("INSERT INTO StatementErrorCallbackTest (randomData) VALUES (?)", ['test']);
        tx.executeSql("THIS STATEMENT WILL FAIL", [],
            function(tx, data) {
                log("FAIL - this statement should have failed");
                finishTest();
            }, function(tx, error) { return false; });
        tx.executeSql("INSERT INTO StatementErrorCallbackTest (randomData) VALUES (?)", ['test1'],
            function(tx, data) { },
            function(tx, error) { log("FAIL - This statement should not have caused an error"); });
    }, function(error) { log("FAIL - The transaction error callback should not have been invoked"); },
    function() { });

    runTransactionExpectedToFail(function(error) { return true; });
    runTransactionExpectedToFail(function(error) { throw "Exception in statement error callback"; return false; });
    runTransactionExpectedToFail(function(error) {});
    runTransactionExpectedToFail(function(error) { return null; });
    runTransactionExpectedToFail(function(error) { return "some string"; });
    runTransactionExpectedToFail(function(error) { return 1234; });
    runTransactionExpectedToFail(function(error) { return {a: 2, b: "abc"}; });
}

</script>
</head>

<body onload="runTest()">
This test confirms that a transaction is immediately rolled back if and only if a statement's error callback throws an exception, returns true, or doesn't return any value.
<pre id="console">
</pre>
</body>

</html>
